edit image

+++ adminpanel/views.py
@@ -211,34 +211,36 @@ def editproduct(request):
         product.product_description = request.POST.get('product_description',False)
         product.category_id = request.POST.get('cate')
 
+        if images1 is not None:
+            format, img1 = images1.split(';base64,')
+            ext = format.split('/')[-1]
+            product_image1 = ContentFile(base64.b64decode(img1), name= product_name + '1.' + ext)
 
-        format, img1 = images1.split(';base64,')
-        ext = format.split('/')[-1]
-        product_image1 = ContentFile(base64.b64decode(img1), name= product_name + '1.' + ext)
+            format, img2 = images2.split(';base64,')
+            ext = format.split('/')[-1]
+            product_image2 = ContentFile(base64.b64decode(img2), name= product_name + '2.' + ext)
 
-        format, img2 = images2.split(';base64,')
-        ext = format.split('/')[-1]
-        product_image2 = ContentFile(base64.b64decode(img2), name= product_name + '2.' + ext)
+            format, img3 = images3.split(';base64,')
+            ext = format.split('/')[-1]
+            product_image3 = ContentFile(base64.b64decode(img3), name= product_name + '3.' + ext)
 
-        format, img3 = images3.split(';base64,')
-        ext = format.split('/')[-1]
-        product_image3 = ContentFile(base64.b64decode(img3), name= product_name + '3.' + ext)
+            if len(request.FILES) != 0:
+                os.remove(product.product_img_1.path)
+                product.product_img_1 = product_image1
+            
+            if len(request.FILES) != 0:
+                os.remove(product.product_img_2.path)
+                product.product_img_2 = product_image2
+            
+            if len(request.FILES) != 0:
+                os.remove(product.product_img_3.path)
+                product.product_img_3 = product_image3
+            product.save()
+            return redirect(allproducts)
 
-        if len(request.FILES) != 0:
-            os.remove(product.product_img_1.path)
-            product.product_img_1 = product_image1
-        
-        if len(request.FILES) != 0:
-            os.remove(product.product_img_2.path)
-            product.product_img_2 = product_image2
-        
-        if len(request.FILES) != 0:
-            os.remove(product.product_img_3.path)
-            product.product_img_3 = product_image3
-
-        product.save()
-
-        return redirect(allproducts)
+        else:
+            product.save()
+            return redirect(allproducts)
     else:
         id=request.GET.get('id')
         product = products.objects.get(id=id)
